fix: Blacklist propagator issues

+++ src/libsyncengine/progress/syncfileitem.h
@@ -62,7 +62,7 @@ class SyncFileItem {
         inline int64_t size() const { return _size; }
         inline void setSize(int64_t newSize) { _size = newSize; }
         inline int progress() const { return _progress; }
-        inline void setProgress(const int newProgress) { _progress = newProgress; }
+        inline void setProgress(const int16_t newProgress) { _progress = newProgress; }
         inline UniqueId operationId() const { return _operationId; }
         inline void setOperationId(const UniqueId newOperationId) { _operationId = newOperationId; }
         inline SyncTime modTime() const { return _modTime; }
@@ -101,7 +101,7 @@ class SyncFileItem {
         CancelType _cancelType{CancelType::None};
         std::string _error;
         int64_t _size{0};
-        int _progress{0}; // %
+        int16_t _progress{0}; // %
         UniqueId _operationId{0};
         SyncTime _modTime{0};
         SyncTime _creationTime{0};
